# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: OneOps Configuration Registry API
  version: "1.0.0"
  description: >
    The system of record for OneOps artifacts (Configuration State Model §6).
    Authority is a maintained projection computed from the dependency graph (M3).
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/admin/policies:
    get:
      summary: List policies
      operationId: listPolicies
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: Policies
          content: { application/json: { schema: { $ref: '#/components/schemas/PolicyList' } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      summary: Create a policy
      operationId: createPolicy
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePolicyRequest' } } }
      responses:
        '201':
          description: Created policy
          content: { application/json: { schema: { $ref: '#/components/schemas/Policy' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/admin/policies/{id}:
    patch:
      summary: Update a policy (enable/disable, condition, action)
      operationId: patchPolicy
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/PatchPolicyRequest' } } }
      responses:
        '200':
          description: Updated policy
          content: { application/json: { schema: { $ref: '#/components/schemas/Policy' } } }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      summary: Delete a policy
      operationId: deletePolicy
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/admin/policies/{id}/executions:
    get:
      summary: List a policy's executions
      operationId: listPolicyExecutions
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Executions
          content: { application/json: { schema: { $ref: '#/components/schemas/PolicyExecutionList' } } }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/admin/policies/{id}/test:
    post:
      summary: Run one synthetic execution of a policy's action
      operationId: testPolicy
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Test result
          content: { application/json: { schema: { $ref: '#/components/schemas/PolicyTestResult' } } }
        '404': { $ref: '#/components/responses/Problem' }

  # Execution timeline (admin) — a read-only operational read model composed from
  # existing persisted data. It participates in no execution.
  /v1/admin/timeline/{eventID}:
    get:
      summary: Execution timeline correlated by audit event id
      operationId: getEventTimeline
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: eventID, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/TimelineFrom' }
        - { $ref: '#/components/parameters/TimelineTo' }
        - { $ref: '#/components/parameters/TimelineComponent' }
        - { $ref: '#/components/parameters/TimelineStatus' }
        - { $ref: '#/components/parameters/TimelineOffset' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Timeline
          content: { application/json: { schema: { $ref: '#/components/schemas/TimelinePage' } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/admin/governance/{id}/timeline:
    get:
      summary: Execution timeline for a configuration object (governance id)
      operationId: getGovernanceTimeline
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/TimelineComponent' }
        - { $ref: '#/components/parameters/TimelineStatus' }
        - { $ref: '#/components/parameters/TimelineOffset' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Timeline
          content: { application/json: { schema: { $ref: '#/components/schemas/TimelinePage' } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/admin/replay/{jobID}/timeline:
    get:
      summary: Execution timeline for a replay job
      operationId: getReplayTimeline
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: jobID, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Timeline
          content: { application/json: { schema: { $ref: '#/components/schemas/TimelinePage' } } }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/admin/policies/{id}/timeline:
    get:
      summary: Execution timeline for a policy execution
      operationId: getPolicyTimeline
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Timeline
          content: { application/json: { schema: { $ref: '#/components/schemas/TimelinePage' } } }
        '404': { $ref: '#/components/responses/Problem' }

  # Compliance & evidence (admin) — read-only. Deterministic evidence composed
  # from existing persisted data; participates in no execution.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items to return (bounded by the server maximum).
      schema: { type: integer, minimum: 1 }
...
    TimelineFrom:
      name: from
      in: query
      required: false
      description: Include entries at or after this RFC3339 time.
      schema: { type: string, format: date-time }
    TimelineTo:
      name: to
      in: query
      required: false
      description: Include entries at or before this RFC3339 time.
      schema: { type: string, format: date-time }
    TimelineComponent:
      name: component
      in: query
      required: false
      schema: { type: string, enum: [governance, audit, webhook, replay, policy] }
    TimelineStatus:
      name: status
      in: query
      required: false
      schema: { type: string }
    TimelineOffset:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0 }
  responses:
    Problem:
      description: RFC 7807 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    PolicyCondition:
      type: object
      properties:
        operations: { type: array, items: { type: string } }
        resources: { type: array, items: { type: string } }
        actors: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: { type: string } }
    PolicyAction:
      type: object
      properties:
        type: { type: string, enum: [http, email, command, notification] }
        config: { type: object }
    Policy:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        condition: { $ref: '#/components/schemas/PolicyCondition' }
        action: { $ref: '#/components/schemas/PolicyAction' }
        max_retries: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PolicyList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Policy' } }
    CreatePolicyRequest:
      type: object
      required: [name, action]
      properties:
        name: { type: string }
        enabled: { type: boolean }
        condition: { $ref: '#/components/schemas/PolicyCondition' }
        action: { $ref: '#/components/schemas/PolicyAction' }
        max_retries: { type: integer }
    PatchPolicyRequest:
      type: object
      properties:
        name: { type: string }
        enabled: { type: boolean }
        condition: { $ref: '#/components/schemas/PolicyCondition' }
        action: { $ref: '#/components/schemas/PolicyAction' }
        max_retries: { type: integer }
    PolicyExecution:
      type: object
      properties:
        id: { type: string }
        policy_id: { type: string }
        event_id: { type: string }
        operation: { type: string }
        cfg_id: { type: string }
        status: { type: string, enum: [pending, running, succeeded, failed, dead_letter] }
        retry_count: { type: integer }
        error: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    PolicyExecutionList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PolicyExecution' } }
    PolicyTestResult:
      type: object
      properties:
        status: { type: string }
        error: { type: string }
    TimelineEntry:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        component: { type: string, enum: [governance, audit, webhook, replay, policy] }
        action: { type: string }
        status: { type: string }
        duration_ms: { type: integer, format: int64 }
        correlation: { type: object, additionalProperties: { type: string } }
        metadata: { type: object, additionalProperties: { type: string } }
    TimelinePage:
      type: object
      properties:
        entries: { type: array, items: { $ref: '#/components/schemas/TimelineEntry' } }
        next_offset: { type: string }
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...